# crates/server/openapi.yaml
...
openapi: 3.0.3
info:
  title: Remote Harness Control Plane API
  version: 0.1.0
  description: |
    Control plane REST API. Human-readable API prose and behavior notes live in
    `docs/API_OVERVIEW.md`; this spec is the machine-readable contract for codegen and review.
servers:
  - url: /
    description: Relative to configured control plane base URL
...
paths:
...
  /workers/tasks/{id}/logs:
    post:
      tags: [workers]
      summary: Append worker log batch for an assigned task
      operationId: postWorkerTaskLogs
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task id (job id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/WorkerLogIngestItem"
      responses:
        "202":
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerLogsAcceptedResponse"
        "400":
          description: Invalid body or timestamps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Task not found or not assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
  /sessions/{id}/logs:
    get:
      tags: [sessions]
      summary: List session log history (cursor pagination)
      operationId: listSessionLogs
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
        - name: job_id
          in: query
          schema:
            type: string
            format: uuid
        - name: level
          in: query
          schema:
            type: string
        - name: last
          in: query
          description: Return the N most recent entries in chronological order (no cursor)
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Paginated log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLogEntries"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
    delete:
      tags: [sessions]
      summary: Delete session logs (optional job scope)
      operationId: deleteSessionLogs
      security:
        - bearerApiKey: []
        - headerApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: job_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "404":
          description: Session or job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: "API key as Bearer token (`Authorization: Bearer <api_key>`)"
    headerApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key in `X-API-Key` header"
  schemas:
...
    StandardError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: object
              additionalProperties: true
...
    WorkerLogIngestItem:
      type: object
      required: [timestamp, level, message, source]
      properties:
        timestamp:
          type: string
          description: ISO 8601
        level:
          type: string
          example: info
        message:
          type: string
        source:
          type: string
          example: worker
    WorkerLogsAcceptedResponse:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
          enum: [true]
    LogEntry:
      type: object
      required:
        - id
        - timestamp
        - level
        - session_id
        - source
        - message
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        level:
          type: string
        session_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
          nullable: true
        worker_id:
          type: string
          nullable: true
        source:
          type: string
        message:
          type: string
    PaginatedLogEntries:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"
        next_cursor:
          type: string
          nullable: true
...